refactor : All Member Entity move to infra
