+++ src/main/java/toolc/yourlist/auth/domain/NonMemberSave.java
@@ -1,0 +1,5 @@
+package toolc.yourlist.auth.domain;
+
+public interface NonMemberSave {
+  NonMember save(NonMember nonMember);
+}
