+++ src/main/java/toolc/yourlist/auth/infra/MemberDomainAdapter.java
@@ -1,7 +1,7 @@
 package toolc.yourlist.auth.infra;
 
 import toolc.yourlist.auth.domain.Member;
-import toolc.yourlist.member.domain.MemberEntity;
+import toolc.yourlist.member.infra.MemberEntity;
 
 public class MemberDomainAdapter {
