+++ src/main/java/toolc/yourlist/auth/infra/NomMemberSaveEntity.java
@@ -1,0 +1,23 @@
+package toolc.yourlist.auth.infra;
+
+import lombok.RequiredArgsConstructor;
+import toolc.yourlist.auth.domain.NonMember;
+import toolc.yourlist.auth.domain.NonMemberSave;
+import toolc.yourlist.auth.domain.PasswordEncoder;
+import toolc.yourlist.member.infra.AllMemberEntity;
+import toolc.yourlist.member.infra.MemberEntity;
+
+@RequiredArgsConstructor
+public class NomMemberSaveEntity implements NonMemberSave {
+
+  private final AllMemberEntity allMemberEntity;
+  private final PasswordEncoder passwordEncoder;
+
+  @Override
+  public NonMember save(NonMember nonMember) {
+    allMemberEntity.save(new MemberEntity(
+      nonMember.uuid(), passwordEncoder.encode(nonMember.uuid()), false));
+    return nonMember;
+  }
+}
+
