--- src/main/java/toolc/yourlist/member/domain/AllMember.java
@@ -1,6 +1,0 @@
-package toolc.yourlist.member.domain;
-
-public interface AllMember {
-  MemberEntity findByLoginId(String loginId);
-  MemberEntity save(MemberEntity member);
-}
