+++ src/main/java/toolc/yourlist/member/domain/MemberRepository.java
@@ -1,5 +1,7 @@
 package toolc.yourlist.member.domain;
 
+import toolc.yourlist.member.infra.MemberEntity;
+
 import java.util.Optional;
 
 public interface MemberRepository {
