+++ src/main/java/toolc/yourlist/member/infra/JpaAllMemberEntity.java
@@ -1,7 +1,6 @@
 package toolc.yourlist.member.infra;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import toolc.yourlist.member.domain.MemberEntity;
 import toolc.yourlist.member.domain.MemberRepository;
 
 public interface JpaAllMemberEntity extends JpaRepository<MemberEntity, Long>, MemberRepository {
